% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cormatrix.R
\name{construct_cormatrix}
\alias{construct_cormatrix}
\title{Construct a correlation matrix by correlating volumes}
\usage{
construct_cormatrix(v, normalize = TRUE, cor.method = "pearson",
  cor.alternative_hypothesis = "two.sided", p.adjust.method = "fdr",
  resample = FALSE, resample_size = NULL, resample_replace = TRUE,
  quiet = FALSE, ...)
}
\arguments{
\item{v}{volume \code{data.frame} or \code{RMINC::anatGetAll} output consisting of individuals as rows and volumes as columns}

\item{normalize}{logical value denoting whether volumes should be normalized (rows summed to 1)}

\item{cor.method}{character string giving the type of correlation coefficient being computed}

\item{p.adjust.method}{character string giving method to use when correcting p values for multiple comparisons}

\item{resample}{logical value denoting whether to resample individuals from the volume \code{data.frame}}

\item{resample_size}{integer value between 3 and the number of rows of the volume \code{data.frame}}

\item{resample_replace}{logical value denoting whether to resample with replacement}

\item{quiet}{logical value denoting whether progress should be printed}
}
\value{
an \code{RSC.cormatrix} object
}
\description{
Creates a correlation matrix by computing pairwise correlations between volumes (column vectors of a \code{data.frame})
}
