% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cormatrix.R
\name{correlate_volumes}
\alias{correlate_volumes}
\title{Construct a correlation matrix by correlating volumes}
\usage{
correlate_volumes(v, cor.method = "pearson", p.adjust.method = "fdr",
  normalize = TRUE, resample = FALSE, resample_N = NULL,
  resample_replace = TRUE, quiet = FALSE)
}
\arguments{
\item{v}{volume \code{data.frame} consisting of individuals as rows and volumes as columns}

\item{normalize}{logical value denoting whether volumes should be normalized (rows summed to 1)}

\item{resample}{logical value denoting whether to resample individuals from the volume \code{data.frame}}

\item{resample_N}{integer value between 3 and the number of rows of the volume \code{data.frame}}

\item{resample_replace}{logical value denoting whether to resample with replacement}

\item{quiet}{logical value denoting whether progress should be printed}

\item{method}{character string giving the type of correlation coefficient being computed}

\item{p.adjust}{character string giving method to use when correcting p values for multiple comparisons}
}
\value{
a cormatrix object
}
\description{
Creates a correlation matrix by computing pairwise correlations between volumes (column vectors of a \code{data.frame})
}

