% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tree.R
\name{hanatFromAnatMatrix}
\alias{hanatFromAnatMatrix}
\title{Construct and populate an anatomical hierarchy from volume data}
\usage{
hanatFromAnatMatrix(anatMatrix, norm.function = NULL, train.indices = NULL,
  hclust.method = "complete", verbose = TRUE, ...)
}
\arguments{
\item{anatMatrix}{A rectangular array of volume data, with structure names as column names.}

\item{norm.function}{A function to normalize the volume data.}

\item{train.indices}{an optional set of indices corresponding to the rows of \code{anatMatrix} indicating the data on which 1) the anatomical hierarchy will be determined, and 2) if \code{norm.function} is set to \code{normByModel}, the data that should be used in the model fitting. If \code{NULL}, then all rows are used in the hierarchy construction.}

\item{hclust.method}{Method to be passed to \code{hclust()}.}

\item{verbose}{Be verbose.}

\item{...}{Further arguments to be passed to the function given by \code{norm.function}.}
}
\value{
An anatomical hierarchy with volume attributes (\code{volumes}, \code{meanVolume}, and if applicable, \code{normVolumes})
}
\description{
Given volume data as a rectangular array, construct an anatomical hierarchy by grouping structures that share volume patterns.
More specifically, a pairwise association matrix is constructed for each pair of structures, where the association is defined as the minimum of the correlation between each of the two structures and the sum of both structures.
In other words, structures are grouped in a way that parent structures (with volume the sum of its children) correlate the most with its children structure. 
The association matrix is then hierarchically clustered, and a data.tree is constructed from the resulting dendrogram. 
Volumes can (and should) be normalized when computing associations.
}
