% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/norm.R
\name{norm_by_model}
\alias{norm_by_model}
\title{Normalization of rectangular array by model fitting}
\usage{
norm_by_model(X, formula, df, train_indices = NULL)
}
\arguments{
\item{X}{A rectangular array, or a vector. If a rectangular array is provided, then the model will be applied to each column separately.}

\item{formula}{a formula describing the model to be fitted, with response variable omitted. If response variable is present, it will be dropped and the appropriate data from \code{X} will be used.}

\item{df}{data frame with same number of rows as \code{X}, containing the predictors.}

\item{train_indices}{an optional set of indices corresponding to the rows of \code{X} indicating the data on which the model should be fit. The same model will be then be applied to the remaining indices. If \code{NULL}, then all rows are used in the model.}
}
\value{
Normalized array or vector of regression residuals.
}
\description{
Fit a model with provided predictors and extract residuals.
}
