% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/norm.R
\name{norm_colregression}
\alias{norm_colregression}
\title{Column sum regression of rectangular array}
\usage{
norm_colregression(X)
}
\arguments{
\item{X}{A rectangular array.}
}
\value{
Normalized array of regression residuals.
}
\description{
Regress out the column sums from elements in a rectangular array. 
Specifically, each row is replaced by residuals of the linear model that predicts row values by sums of columns.
This is similar to dividing by column sums (i.e. \code{norm_colsum()}) but also takes the differences in slopes (w.r.t. to the column sums) of each row into account.
}
