% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/norm.R
\name{norm_fullsum}
\alias{norm_fullsum}
\title{Total sum normalization of rectangular array}
\usage{
norm_fullsum(X)
}
\arguments{
\item{X}{A rectangular array.}
}
\value{
Normalized array with all elements summing to 1.
}
\description{
Normalize a rectangular array by dividing each element by the sum of all elements.
}
