% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/norm.R
\name{norm_successive}
\alias{norm_successive}
\title{Successive normalization of rectangular array}
\usage{
norm_successive(X, type = "F", tol = 1e-08, verbose = TRUE, na.set = 0)
}
\arguments{
\item{X}{A rectangular array.}

\item{type}{Normalization method. This argument is passed as \code{type} to the \code{norm()}.}

\item{tol}{Tolerance for convergence.}

\item{verbose}{Be verbose .}

\item{na.set}{\code{norm()} requires that no elements in the rectangular array be \code{NA}. Replace \code{NA}s with this value.}
}
\value{
Normalized array.
}
\description{
A method of successively normalizing both rows and columns of a matrix, a la Brad Efron and further described by Olshen et al.
See for more details: https://www.ncbi.nlm.nih.gov/pmc/articles/PMC2868388/ . 
Note that the proof on convergence contained within aforementioned is flawed according to the 2013 paper published by the same authors.
}
