% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tree_inference.R
\name{posterior_inference}
\alias{posterior_inference}
\title{Correlation coefficient inference via Bayesian updating on hierarchical anatomy}
\usage{
posterior_inference(hanat, struc_i, struc_j, indices = NULL,
  p2r.table = NULL, thres_cor = 0, precision_cor = NULL,
  precision_sampling = 100, volattr = "normVolumes")
}
\arguments{
\item{hanat}{Hierarchical anatomy tree giving relationship between structures and their parents.}

\item{struc_i}{Structure name (must be contained as \code{name} attribute in some node of \code{hanat}).}

\item{struc_j}{Structure name (must be contained as \code{name} attribute in some node of \code{hanat}).}

\item{indices}{Optional set of indices that indicate which volumes/observations to use in computing the posterior data. This is useful if you store all volumes across different experimental groups in the tree, and want to separately compute the posterior data in different groups.}

\item{p2r.table}{Optional table (output of \code{probability_to_rho_table()}) that maps probabilities to correlations. Provide this is you want a posterior correlation estimate.}

\item{thres_cor}{Threshold correlation from which to calculate the probability. Default is 0.}

\item{precision_cor}{Optional parameter to tune the width of the distribution of the correlation coefficient p(r). By default this is equal to the number of indices (if provided) or volumes/observations (if indices are not provided).}

\item{precision_sampling}{Optional parameter to tune the quality of the estimated distributions, which are calculated by sampling a number of times given by this parameter.}

\item{volattr}{Attribute name in the anatomy tree \code{hanat} that contains the volume data from which correlations are computed.}
}
\value{
a list of posterior data including the probability that the posterior correlation is above the input threshold, and (if \code{p2r.table} is provided), a posterior correlation.
}
\description{
Correlation matrices constructed from small samples (e.g. n=10) drawn from the same population are generally not very robust, in the sense that across multiple draws, each correlation coefficient varies quite a bit.
This is a statistical issue: the distribution of correlation coefficients, when sampling from a population with known correlation, is very wide when the sample size is small. 
For example, when sampling the correlation from 10 observations repeatedly drawn from a population correlation of 0.5, the 95% of the distribution p(r) lies between -0.15 and 0.86. 
When the number of observations increases to 100, the interval becomes much narrower, around 0.33 to 0.63. 
Typically biological studies are not fortunate to have sample sizes that large. Brain structures are constrained in their correlation patterns however (e.g. two structures that develop from a parent structure are both likely to have similar correlations in volumes to the parent structure); 
and exploiting such constraints might provide better estimates of correlations between structural properties than the raw sample correlation coefficient.
This function uses a Bayesian framework and exploits the hierarchical structure of anatomy in order to better estimate the correlation coefficient between two given structures. 
Briefly, the probability that the correlation between two structures is greater than a threshold value is inferred by recursively computing correlations between their parent structures, and updating this probability (modeled as a beta distribution).
Optionally, a mapping from probability to correlation can be also specified to convert the final probability to a posterior correlation.
}
